NBT serialization finished up, merged telemetry features with CaliforniaDemise's version, rejected the fear of variable-size buffers

+++ src/main/java/com/starl0stgaming/gregicalitystarbound/api/telemetry/network/packet/TelemetryPacket.java
@@ -38,7 +38,7 @@ public class TelemetryPacket {
         return destinationID;
     }
 
-    public void setDestinationID(long destinationID) {
+    public void setDestinationID(int destinationID) {
         this.destinationID = destinationID;
     }
